refactor: move memory DAOs into memory package
