+++ server/src/main/java/dataaccess/AuthDAO.java
@@ -4,7 +4,7 @@ import model.AuthData;
 
 public interface AuthDAO {
     void createAuth(AuthData authData) throws DataAccessException;
-    AuthData getAuth() throws DataAccessException;
+    AuthData getAuth(String authData) throws DataAccessException;
     void deleteAuth(String auth) throws DataAccessException;
 
     void clear() throws DataAccessException;
