+++ server/src/main/java/dataaccess/memory/MemoryGameDAO.java
@@ -1,6 +1,8 @@
-package dataaccess;
+package dataaccess.memory;
 
 import chess.ChessGame;
+import dataaccess.DataAccessException;
+import dataaccess.GameDAO;
 import model.GameData;
 
 import java.util.Collection;
